Call `annotateHoleData()` from top-level Pianolatron component
